added indices

+++ Valis/VBO.cu
@@ -1,12 +1,14 @@
 #include "VBO.cuh"
 
+#include <stdint.h>
+
 VBO::VBO(void* data, size_t size, BufferedObjectUsage usage)
 {
 	glGenBuffers(1, &handle);
 
 	glBindBuffer(GL_ARRAY_BUFFER, handle);
 
-	glBufferData(GL_ARRAY_BUFFER, size * sizeof(float), data, usage);
+	glBufferData(GL_ARRAY_BUFFER, size * sizeof(uint32_t), data, usage);
 
 	glBindBuffer(GL_ARRAY_BUFFER, 0);
 }
@@ -17,7 +19,7 @@ VBO::VBO(size_t size, BufferedObjectUsage usage)
 
 	glBindBuffer(GL_ARRAY_BUFFER, handle);
 
-	glBufferData(GL_ARRAY_BUFFER, size * sizeof(float), NULL, usage);
+	glBufferData(GL_ARRAY_BUFFER, size * sizeof(uint32_t), NULL, usage);
 
 	glBindBuffer(GL_ARRAY_BUFFER, 0);
 }
